Move some templates to header, using template in send_response
Also in main, stop looking for arguments when we reach the -- flag.
